- fixed: Dehacked's 'Speed' value should be treated as signed when being assigned to a double variable.

+++ src/d_dehacked.cpp
@@ -921,7 +921,7 @@ static int PatchThing (int thingy)
 		{
 			if (stricmp (Line1, "Speed") == 0)
 			{
-				info->Speed = val;	// handle fixed point later.
+				info->Speed = (signed long)val;	// handle fixed point later.
 			}
 			else if (stricmp (Line1, "Width") == 0)
 			{
